Adding Checker, Bug Fix Problem Moderator, Add requesting for Moderator

+++ app/Http/Controllers/Administration/Problem/ModeratorController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Services\Problem\ProblemService;
 use App\Models\User;
+use App\Models\ModeratorRequest;
 
 class ModeratorController extends Controller
 {
@@ -29,7 +30,10 @@ class ModeratorController extends Controller
 
     public function addModerator()
     {
-
+        if($this->problemData->authUserRole != "owner")
+        {
+            abort(401,"You Can Not Add Moderator. Only Problem owner can add moderator");
+        }
         $this->problemData->moderator()->attach(request()->userId,[
             'role' => 'moderator',
             'is_accepted' => 0
@@ -41,10 +45,34 @@ class ModeratorController extends Controller
 
     public function deleteModerator()
     {
+        if($this->problemData->authUserRole != "owner")
+        {
+            abort(401,"You Can Not Add Moderator");
+        }
+        $user = $this->problemData->moderator()->where('user_id',request()->userId)->firstOrFail();
+
+        if($user->pivot->role == "owner")
+        {
+            abort(401,"You Can Not Be Deleted");
+        }
         $this->problemData->moderator()->detach(request()->userId);
         return response()->json([
             'message' => "Moderator Added Successfully",
         ]);
+    }
+
+    public function leaveModerator()
+    {
+        if($this->problemData->authUserRole == "owner")
+        {
+            abort(401,"You Can Not Leave From this Problem");
+        }
+        $this->problemData->moderator()->detach(auth()->user()->id);
+        return response()->json([
+            'message' => "You Leave From {$this->problemData->name}",
+            'url' => route('administration.problems'),
+        ]);
+
     }
 
     public function cancelModeratorRequest()
@@ -53,6 +81,7 @@ class ModeratorController extends Controller
         $this->problemData->moderator()->detach(auth()->user()->id);
         return response()->json([
             'message' => "Moderator Detach Successfully",
+            'url' => route('administration.problems')
         ]);
         
 
@@ -65,5 +94,27 @@ class ModeratorController extends Controller
         return response()->json([
             'message' => "Moderator accept successfully",
         ]);
+    }
+
+    public function requestForModerator()
+    {
+        if (!auth()->check()) {
+            abort(401,"You need to login your account");
+        }
+        if(!auth()->user()->moderatorRequest)
+        {
+            ModeratorRequest::create([
+                'user_id' => auth()->user()->id,
+                'type' => 30
+            ]);
+        }
+
+        // dd(User::has('moderatorRequest')->get());
+        
+        return response()->json([
+            'message' => "Your Request Is Sent"
+        ]);
+
+
     }
 }
